EntityManager: Architecture overview
This document intended to describe system architecture as we would like it to be,
but not actual implementation state. Main goal is to define system architecture
and EM role in there that would be clear and common for everyone interested in
using EM. When we accept this implementation proposal, we should keep in mind
this design when refactoring code and/or developing new features.

Signed-off-by: Andrei Kartashev <a.kartashev@yadro.com>
Change-Id: Ib9033e1bcfae3a76456c22234c58193c09959cb3