systemd: new obmc-chassis-emergency-poweroff@.target
When firmware detects a critical power or thermal error (i.e. something
that could potentially cause physical damage to the system), it's
critical that it has a way to quickly turn off the system and notify
the user.

This new obmc-chassis-emergency-poweroff@.target is that way. It will utilize
both the obmc-chassis-hard-poweroff@.target(to power off quickly) and the
obmc-host-quiesce@.target(to notify the user that the host is no longer
running and the system is in an error state).

This target will be special in that no automated recovery will occur
when it is started. It is assumed that if the target is started, then user
intervention is required to diagnose and repair the system.

Signed-off-by: Andrew Geissler <geissonator@yahoo.com>
Change-Id: Ib1e3a653db7ab9b348158410c40d33215ff90326