Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > This is pulled from the one I wrote?  There is omissions in that that we should have in a "generic" clang-tidy file.
> 
> Yes. 
> 
> > There is omissions in that that we should have in a "generic" clang-tidy file.
> 
> Like?

Cert checks that bmcweb had an errant example of, performance changes that should be enabled, but some bmcweb patterns were used a lot, and a couple others I'd have to check the list for that I haven't looked at off the top of my head.

What's the driving force for getting this checked in now before we've done a couple repos?  Is there effort to try to use this in other places?  My plan was to get a couple of repos going with configs that we're happy with, then try to standardize.  I was also hoping to get clang-11 into the build CI before we did that, as it adds checks that would be useful.

Patch-set: 1