Update patch set 1
Patch Set 1:

(1 comment)

> Patch Set 1:
> 
> > Is there effort to try to use this in other places?  
> 
> I've been playing around with this in openpower-proc-control.

Cool.  I will try to consolidate what I would like to see a "stock" config to be for most repos.  Might be tomorrow.

My assumption would be that we would start with a superset of checks in the docs repo, and the individual repos could remove from it where it would be too hard to transition patterns, and we can ratchet up the checks over time.

Patch-set: 1