designs/mctp: Move MCTP designs to a subdirectory, making way for kernel approach
We've had long-term plans for a kernel approach to MCTP, but have
documented the current userspace-based implementation here.

This change moves the MCTP design to a new subdirectory (mctp/), and
splits the design into a common overview (mctp.md) and the current
userspace-based design (mctp-userspace.md).

This allows us to introduce a kernel design as a future commit, and
share the overview amongst documents.

Signed-off-by: Jeremy Kerr <jk@codeconstruct.com.au>
Change-Id: I834bce8bd3b6a3a89e00e0b3ff9cd77014d485a1