Update patch set 1
Patch Set 1:

(4 comments)

I reviewed this originally thinking a lot of stuff that wasn't new was.  It would help a lot to split up the formatting changes into a commit preceding adding the new commands.

Overall, this whole file is technically undefined behavior in Redfish, as nothing is following the hypermedia links which have no guarantees of consistency in Redfish.  I'm not a huge fan of us documenting that kind of thing, and would much rather we point to redfishtool, redfish tacklebox, or pick-your-favorite-redfish-client scripts, which parse the hypermedia endpoints properly.  With that said, your patchset only makes this slightly worse, so I'm conflicted on that.  It's a lot more to clean up when someone goes through and does it properly, but in the meantime it's good documentation.

At one point, Gunnar and I talked about moving this file into bmcweb, as a means to capture peoples Tested statements for the long run, and hopefully start amassing useful scripts for things when people check in new features.  That's a ways out though.

Patch-set: 1
Reviewer: Gerrit User 1000153 <1000153@adbd0a64-1f21-4d83-b585-671fe73cb6e4>