Update patch set 5
Patch Set 5:

(12 comments)

Overall this looks good to me, and ok to clarify it in separate commit too. 
Note: 
Following things needs to be planned as well. 

1. Support for SPDM & secure Message type.
i.e. SPDM daemon can handle the message types, but it would be better for the SPDM to configure the session details to the kernel, such that kernel can use the same for further transmit / receive of the secure MCTP messages (This will ease up the load). 

2. NULL EID clarifications or MCTP Message communication with Null EID communication (Don't find any issue, as long as the physical interface device address is clarified - you already provided a way).

Patch-set: 5
Reviewer: Gerrit User 1000168 <1000168@adbd0a64-1f21-4d83-b585-671fe73cb6e4>