Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review+1
> 
> Can we please merge this and future iteration can be submitted in the form of patches to this document?

We’ve been waiting over a year for a new patch set as promised by the author. If I merge this I have no confidence that the design will ever be agreed to at this point. It seems like it is not a priority for the original author. Joel, is this something you or someone else can take over?

Patch-set: 2
Reviewer: Gerrit User 1000000 <1000000@adbd0a64-1f21-4d83-b585-671fe73cb6e4>
Label: Ok-To-Test=0