Update patch set 12
Patch Set 12:

(4 comments)

Thanks for the reviews, Richard, I appreciate the time! Andrew has covered most of the comments, just a few replies inline.

If there's any tweaks we need as a result, I think the best approach is for a future change, as there's no fundamental issues as yet.

Patch-set: 12