Create template application
This application should serve as a template for how to create and launch
a simple dbus server application.

It is relatively simple app that:
Requests a name on dbus
Registers a simple interface at /xyz/openbmc_project/template_app with
one property and one method.

It also creates unit tests to test the various components.

The .clang-format and .clang-tidy files are pulled from the normal
template locations.  Going forward, this will be where they live.

This is not the final location for these, they will be checked into a
new repo.

https://lore.kernel.org/openbmc/CACWQX80G75i+s8Vcro64uEyvcZn1Dc60y-coS6GqrvkJo59Kqg@mail.gmail.com/

DO NOT MERGE

TODO:
Still needs phosphor-dbus-interfaces example.

Signed-off-by: Ed Tanous <edtanous@google.com>
Change-Id: Id138ccffa18e38ec43cef31907b396079536ea6e