Update patch set 2
Patch Set 2:

Ed, is this the right place for this doc? I would expect it to live in dbus-sensors, but I see the advantage of centralized docs. If so, reference this doc in the readme and source for ExternalSensor.

Patch-set: 2
Reviewer: Gerrit User 1000502 <1000502@adbd0a64-1f21-4d83-b585-671fe73cb6e4>