Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Ed, is this the right place for this doc?

Yes.

> I would expect it to live in dbus-sensors, but I see the advantage of centralized docs. If so, reference this doc in the readme and source for ExternalSensor.

Unfortunately, that's not how that works today.  I can't say I like centralizing the documentation, and I agree with you it would be better if distributed, but for the moment consistency is a greater user-win than ease of entry.

If you want to propose a change to the way we handle design reviews, this commit isn't the place to do it;  I would start with a mail to the mailing list, and push some patchsets to start moving the design docs to their appropriate "homes".

Patch-set: 2
Reviewer: Gerrit User 1000153 <1000153@adbd0a64-1f21-4d83-b585-671fe73cb6e4>