Update patch set 6
Patch Set 6:

This is a great start. Until phosphor-webui  has been officially deprecated, there are still people using it, do we want to just add a section for webui-vue? I realize that the information removed is not needed for webui-vue and I personally would expect to find specific documentation about cloning, installing, and running on the code repo, but should we leave that info for people interested in phosphor-webui? I appreciate how you linked to the contributing section and think that is a best practice. I am not used to going to a markdown file outside of the repo to find info on installation and contribution.

It might be nice to add a section above contributing that suggests why this GUI was created in the same pattern you used for contributing and point to the sections in https://github.com/openbmc/webui-vue/blob/master/README.md

Patch-set: 6
Reviewer: Gerrit User 1000281 <1000281@adbd0a64-1f21-4d83-b585-671fe73cb6e4>