Update patch set 1
Patch Set 1:

(20 comments)

Thanks for the encouragement and the quick review.

As noted in the comments, the intent of this design is to address only: custom roles, OEM privileges, and role-to-privilege mapping.  It is not intended to address customizable privilege registries (the operation-to-privilege mapping).  I apologize if the design did not make that clear: I will rewrite to clarify that.

As you pointed out, being able to customize the role-to-privilege map without the ability customize the operation-to-privilege mapping is mostly useless.  Nevertheless, I see it as an important building block deserving a separate design.  What do you think?

--

I believe my comments address your 4 main comments:
> 1. Building multiple times per machine...

This one is expensive to solve in general.  See my comments.  When this is implemented, I intend to use it right away to define my OemIBMServiceAgent role and OemIBMPerformService privilege for selected machines.  So we'll get there.

> 2. [Mention past attempts to customize] runtime privilege registry.

I am not aware of any attempted designs (other than my earlier attempts) that limit their scope to customized roles and privileges, but omit the privilege registry (but I will look again).

> 3. [S]olve the per-route privilege mapping.

This design avoids the "per-route privilege mapping" problem by avoiding the "customized privilege registry" problem.  As I explained above, I think separating these designs is valid.

> 4. [Don't] require every custom system to write out the complete role targets.

Agreed.  I proposed a better approach in the comments.

Patch-set: 1