Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Jason - can you explain your proposal that you proposed in the GUI Work group meeting ?

Yep.  My thought was that we already treat all three params as a string in the code, so calling them a number in the Message Registry doesn't really change anything.
        std::vector<std::string> messageArgs = {
            std::to_string(bootIndex), timeOffsetStr.str(), hexCode.str()};

If we just change at least the last param to a string (if not all three):
MessageEntry{"BIOSPOSTCode",
             {
                 "BIOS Power-On Self-Test Code received",
                 "Boot Count: %1: TS Offset: %2; POST Code: %3",
                 "OK",
                 "OK",
                 3,
                 {"number", "number", "string"},
                 "None.",
             }},

That leaves it open for whoever is sending the Message to decide what's in the param.  The existing implementation will still work since the POST Code number is just a string that will still be displayed, and the new implementation can just display the ASCII code instead of the number.  That way both implementations can use the same Message.

Patch-set: 1
Reviewer: Gerrit User 1000276 <1000276@adbd0a64-1f21-4d83-b585-671fe73cb6e4>