Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> The OpenBMC security working group discussed some aspects of this design today.  Notes are in the minutes for 2021-05-12.  Please take a look, and address comments to the appropriate reviews.
> 
> https://docs.google.com/document/d/1b7x9BaxsfcukQDqbvZsU2ehMq4xoJRQvLxxsDUWmAOI

With respect, no.

If the security working group wants to have an impact beyond its meetings it needs to do the leg-work. This means engaging with the relevant developers through the usual means, where the usual means in this case are leaving review comments against the relevant sections of patches in Gerrit.

I'm not going to put any more effort into addressing the working group's concerns on the proposal than they're willing to put in to the review process. If it's too much effort to leave appropriate reviews then I can only assume that the problems are not that concerning and therefore don't warrant any action on my part.

That said, I'm cautious of giving the impression here that I might be trying to avoid addressing security concerns. I want to be clear that it isn't the case, and that the document already discusses some security issues exposed by the proposed feature. Again, if this discussion needs work, then please leave comments indicating what outcomes you're after.

Patch-set: 5