Update patch set 4
Patch Set 4:

> Patch Set 1:
> 
> I see in your proposal, just presence information is detected/checked. No other inventory data.
> I see we already have phosphor-nvme which supports to read inventory information and check GPIO presence but no Redfish support.
> How's about supporting a framework (or dbus interface) so that services like phosphor-nvme can update information to so that the information can display in Redfish SimpleStorage?

Sorry for so long to reply you

The existing bmcweb already has the storage scheme, which can display the information of the Storage Device, it used the dbus interface , xyz.openbmc_project.Inventory.Item.Drive.
Considering the duplication of SimpleStorage and Storage, the design proposal for simpletorage has been abandoned.

Patch-set: 4