Update patch set 2
Patch Set 2:

(1 comment)

The card has the simplest logic, a spst switch that is open or closed.   The desire is for the simplest interaction with the user not the simplest code.   The rule is if the switch is moved from the position when the BMC was last booted then reset
   There is no memory in the switch.   There is no "and now put it back" double operation.

Patch-set: 2
Reviewer: Gerrit User 1000021 <1000021@adbd0a64-1f21-4d83-b585-671fe73cb6e4>