Update patch set 1
Patch Set 1:

(45 comments)

Lots of comments;  Overall I think this would be useful, but am I little worried about the details, and the amount of scope this one document is taking on.  I'm happy to discuss in individual patchsets as this is incrementally added, but please ensure that as you're doing this work, you're separating the new code from the refactored old code, as I suspect there will be a lot of conflicts.

Patch-set: 1
Reviewer: Gerrit User 1000153 <1000153@adbd0a64-1f21-4d83-b585-671fe73cb6e4>