Update patch set 1
Patch Set 1:

As an aside, this is one of those times where I wish we'd adopted the go-redfish (now called sailfish) model of tracking a "real" resource tree within the redfish implementation.  That would solve a lot of issues that are going to pop up with this design;  Part of me wonders if we could use the existing code to build the resource tree model directly, and deprecate the route handlers.

Patch-set: 1
Reviewer: Gerrit User 1000153 <1000153@adbd0a64-1f21-4d83-b585-671fe73cb6e4>