Update patch set 4
Patch Set 4:

I'm having a little trouble getting back to this.  The outstanding item is writing a phosphor-dbus-interfaces version of the daemon, which I haven't really had a chance for.  Do we want to keep this in review while we wait for that, or do we want to merge this into a new repo, and treat that as a second patchset?  Happy either way, but I think as-is this example is still useful, and I've pointed a lot of people to this review since I first created it.

Patch-set: 4