Update patch set 1
Patch Set 1:

I am open to any and all suggestion for the naming convention and placement in the document. The gist of it is that this GPIO indicates a faulty voltage regulator that regulates the standby voltage. The regulator may or may not be a module (the m in vrm). Other platforms are not prohibited from implementing a similar design.

Patch-set: 1