Update patch set 2
Patch Set 2:

> Changing the name will also require changing the associated source code. The easiest thing for me to do is to keep the name as-is and move the definition to the non-POWER-specific, Special, section of this document.

Please do not say things like this.

Refactoring code is cheap; the cognitive overhead of poor and inconsistent names is not.  If you are unable to do a find and replace on your code base to fix up a name you chose to use before coming to consensus on an appropriate name, we need to fix that.

Patch-set: 2
Reviewer: Gerrit User 1000000 <1000000@adbd0a64-1f21-4d83-b585-671fe73cb6e4>