Update patch set 1
Patch Set 1: Code-Review-1

I appreciate the TSC members giving a +1 on this.  That gives good indication that conceptually the CLA updates are acceptable at a project level.  There is still work to be done on the mechanics of how these CLAs are stored in git, so that is what is holding off merge.

Patch-set: 1
Reviewer: Gerrit User 1000000 <1000000@adbd0a64-1f21-4d83-b585-671fe73cb6e4>
Label: Code-Review=-1