Update patch set 2
Patch Set 2:

> Patrick what would you like the name to be changed to?

I've not given any opinion on the naming thus far with this commit and I don't necessarily have the domain expertise to do so.  Added Matt.

Matt / Ben, reading shortly on this it feels to me like this may be something we want to have a cohesive design on fault/pgood indicators to the BMC, like we have with some of the other sections (led, buttons, etc.).  Should we define a section and schema for power failure and status indications?  Or is this really a one-off?

Patch-set: 2
Reviewer: Gerrit User 1000000 <1000000@adbd0a64-1f21-4d83-b585-671fe73cb6e4>