Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patrick what would you like the name to be changed to?
> 
> I've not given any opinion on the naming thus far with this commit and I don't necessarily have the domain expertise to do so.  Added Matt.
> 
> Matt / Ben, reading shortly on this it feels to me like this may be something we want to have a cohesive design on fault/pgood indicators to the BMC, like we have with some of the other sections (led, buttons, etc.).  Should we define a section and schema for power failure and status indications?  Or is this really a one-off?


IBM wires all of its other VRM faults to a power sequencer device, and the code that deals with that is outside the scope of this change.  This one was wired to the BMC because a fault here would cause a bmc power loss/reboot.  Now in state manager when it detects it's coming back up when main power was previously up it can check this gpio and report a regulator error as opposed to a power blackout error: https://gerrit.openbmc-project.xyz/c/openbmc/phosphor-state-manager/+/44860

I have no idea if this is common practice or not, though I imagine not.

Patch-set: 2
Reviewer: Gerrit User 1000008 <1000008@adbd0a64-1f21-4d83-b585-671fe73cb6e4>