Update patch set 1
Patch Set 1:

> I'm assuming this leaves the already broken up string literals alone?

Based on a simple experiment correct.  It does not appear that clang-tidy will "unbreak" your string literals.  It will only stop doing it in the future.

--

I personally prefer to not have anything over 80 characters, but I understand why the consensus will not be in my favor on this one.  Assuming I am correct that the consensus will be on the side of this change, I'll merge it in a few days.

Patch-set: 1
Reviewer: Gerrit User 1000000 <1000000@adbd0a64-1f21-4d83-b585-671fe73cb6e4>