bmc-reset: ensure recovery when host unresponsive
This new feature will ensure an error is logged and the host is put in
quiesce state in situations where the host was attempting to boot before
a BMC reboot, and crashed while the BMC was going through its reboot.
Going to the quiesce state will kick in whatever recovery has been
defined for the system.

There are windows during the boot of the host where it requires the BMC
be available, and if not, it will crash itself. A recovery and clean
reboot is much simpler then handling all of the different corner cases
that can occur in this scenario.

Change-Id: Id1f2f326d08d4d77a38fffb0cfd2227fb91453e1
Signed-off-by: Andrew Geissler <geissonator@yahoo.com>