bmc-service-failure: handling a service in the fail state
This commit adds additional details and requirements on how to handle an
OpenBMC systemd service that enters the fail state.

The service should not just silently fail, it needs to be reported to
the system owner and appropriate debug data needs to be generated so
root cause can be identified.

Change-Id: I0527213bcf2c5886554d02c238cb963cc00f0f87
Signed-off-by: Andrew Geissler <geissonator@yahoo.com>