device-tree-gpio-naming: Add POWER vref gpio
The vref gpio on POWER systems is used by the BMC to tell the HW if a
system requires 2 or 4 power supplies. If enough power supplies fail so
that the system no longer has redundancy (no longer N+1), the hardware
will signal to the Onboard Chip Controller that the system may be
oversubscribed, and performance may need to be reduced so the system can
maintain its powered on state.

Change-Id: I9d14b869765df0755262f6e5f14c0fed03736640
Signed-off-by: Adriana Kobylak <anoo@us.ibm.com>