virtual-media: add support for asynchronicity
Sometimes mounting and unmounting can be time consuming. Low performing
platforms mount remote resources in time exceeding 25 seconds of DBus
timeout. In order to address that, support for asynchronous DBus and
Redfish calls has to be defined by design.

This is not an issue on faster platforms so it was not considered
before.

Also Redfish schemas example has been extended with actions.

Change-Id: I651ac17ae5091e12aa78cecedf6a9f4a2c7421d8
Signed-off-by: Przemyslaw Czarnowski <przemyslaw.hawrylewicz.czarnowski@intel.com>