Update patch set 1
Patch Set 1:

Please take a look at https://github.com/openbmc/docs/blob/master/CONTRIBUTING.md#inclusive-naming. We as a project have tried to deprecate certain terms (i.e. whitelist == allowlist).

One issue we have with our AST2600 and eMMC usage is that we preserve /etc and /var across firmware updates. There are times where we'd like to do a mini factory reset, where we keep the network settings and maybe a few other things, but clear most everything else. I wonder if this design is something we could build on for that type of function. This is mostly a "clean-up-the-lab-system" type function.

Patch-set: 1
CC: Gerrit User 1000004 <1000004@adbd0a64-1f21-4d83-b585-671fe73cb6e4>