Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Please take a look at https://github.com/openbmc/docs/blob/master/CONTRIBUTING.md#inclusive-naming. We as a project have tried to deprecate certain terms (i.e. whitelist == allowlist).
> 
> One issue we have with our AST2600 and eMMC usage is that we preserve /etc and /var across firmware updates. There are times where we'd like to do a mini factory reset, where we keep the network settings and maybe a few other things, but clear most everything else. I wonder if this design is something we could build on for that type of function. This is mostly a "clean-up-the-lab-system" type function.

Hello,

I pushed a change for initrdscripts (which this design depends on) to remove the deprecated terminology.
https://gerrit.openbmc-project.xyz/c/openbmc/openbmc/+/47200
For AST2600, /etc and /var are only preserved when flashing the normal `mtd.tar` images but not the `mtd.all.tar` images, in which case preservation is handled on a per-file basis by the update script. This design is to allow managing the configuration of this update script, which may be more desirable than preserving the entire rwfs for some use-cases.

Patch-set: 1