power-recovery: bmc and system recovery paths
Provide a mechanism for the power recovery software to not run in
situations where a user has manually intervened to try and recover the
system.

For example, if a system is unresponsive and a user manually resets
it in some way (for example via a pin hole reset), then do not run
the power recovery logic. This allows the user to manually work with
the system to debug and not worry about it automatically doing things
on them.

Change-Id: I3bead46df5d7ad4344d47affc066c7e36379e0db
Signed-off-by: Andrew Geissler <geissonator@yahoo.com>