Update patch set 1
Patch Set 1:

(2 comments)

If a repo is not updated to use meson and the recipe gets deleted, but the repo is necessary for most of the systems, who will be responsible to update the repo to use meson? Will there be some updates to the maintainers of the repo?

Patch-set: 1
Reviewer: Gerrit User 1000044 <1000044@adbd0a64-1f21-4d83-b585-671fe73cb6e4>