reset-cause: define new pinhole reset gpio
The pinhole reset cause will be utilized by BMC firmware to know
when it has been reset due to a user initiated pinhole reset. This
is commonly done in error scenarios where the BMC is hanging or
otherwise unresponsive.

Initially I put this under the Buttons section but as it's really more
of a reason that a pinhole reset has occurred, I went with a new
section. The BMC is not involved with the actual pinhole reset hardware
because the whole purpose of it is to trigger a BMC reset when the BMC
is unresponsive.

The design doc on this function is at:
  https://gerrit.openbmc-project.xyz/c/openbmc/docs/+/47816

Change-Id: I9fb4e5909a279dc588815bb13ff1550c70f867ab
Signed-off-by: Andrew Geissler <geissonator@yahoo.com>