indicators: define new pin hole reset gpio
The pin hole reset indicator will be utilized by BMC firmware to know
when it has been reset due to a user initiated pin hole reset. This
is commonly done in error scenarios where the BMC is hanging or
otherwise unresponsive.

Initially I put this under the Buttons section but as it's really more
of an indication that a pin hole reset has occurred I went with a new
section. The BMC is not involved with the actual pin hole reset hardware
because the whole purpose of it is to trigger a BMC reset when the BMC
is unresponsive.

Change-Id: I9fb4e5909a279dc588815bb13ff1550c70f867ab
Signed-off-by: Andrew Geissler <geissonator@yahoo.com>