CLA: Update links to CCLA and ICLA
The TSC approved new CLAs templates:
https://gerrit.openbmc-project.xyz/c/openbmc/docs/+/43587 that avoid
requiring contributors to divulge sensitive personal information.

The old templates were hosted on a GitHub asset server attached to a
comment in openbmc/openbmc#3046.  This was a poor choice for a number of
reasons - the main one being it requires a GitHub support ticket to get
the assets taken down.

For the moment, the new templates are hosted in the same Google drive as
the filled out CLAs.

This patch updates any links to the templates in the docs repository.

Change-Id: I846451e56bb0b0d5206e74084c916741535ebb43
Signed-off-by: Brad Bishop <bradleyb@fuzziesquirrel.com>