Update patch set 1
Patch Set 1:

> I don't think this document is the place to question the design choices made in the board layout.

This is a rather simplistic description of the previous discussion.

There are two arcs of discussion here:

1. Is 'battery-voltage-read-enable' sufficiently descriptive of the function?
2. Does the function being described warrant documentation in a generic section or is this something peculiarly implemented by a particular board(s)?

For 1, I don't believe so.  I'd, at a minimum, like to see 'rtc' in the name.

For 2, that is probably what you perceived as "question the design choices made in the board layout".  If this is not a reasonable design, based on the hardware being used, my opinion is that we should not document it in a generic way so that others aren't mislead into thinking it is something that provides significant benefit.  If this is not a best practice (or even good practice) we should be clear that that is the case.

We have had discussions before about if this document is intended to be "document every GPIO that everyone has ever used" or "document the patterns of GPIOs we expect to be able to reuse in the code-base".  We've tended to settle on the latter.  We could certainly explore an 'OEM' nomenclature here as well, which we have sort of done via the "POWER Specific GPIOs" section.

Currently the Requirements section states what I think matches the "document the patterns" philosophy:
> Ensure common function GPIO's within OpenBMC use the same naming convention

Patch-set: 1
Reviewer: Gerrit User 1000000 <1000000@adbd0a64-1f21-4d83-b585-671fe73cb6e4>