Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+1
> 
> Sounds good to me. Looking at https://clang.llvm.org/docs/ClangFormatStyleOptions.html, it's not totally clear to me how setting these two causes what we want here but clang doc isn't always all that clear.

`DeriveLineEnding: false`
> Analyze the formatted file for the most used line ending (\r\n or \n). UseCRLF is only used as a fallback if none can be derived.

This turns off automatic detection of unix vs dos mode.

`UseCRLF: false`
> Use \r\n instead of \n for line breaks. Also used as fallback if DeriveLineEnding is true.

Since DeriveLineEnding is off, this is how we are forcing to unix mode (LF) instead of dos mode (CRLF).  

If we left DeriveLineEnding on, then UseCRLF would only be used in the case where clang-format couldn't figure it out, such as if there were a mixture in the file.  We don't want to allow the authors to pick their own, even if they are consistent, so we turn off detecting and force a particular one.

Patch-set: 1