development: remove tutorial on adding new system
As noted in openbmc/docs#68, the tutorial to add a new system is
becoming a bit outdated. It was never a full end to end doc to begin
with and it doesn't really fit well within a tutorial anyway.

Ideally we as a community get a full blown document describing all the
different areas a user needs to update to add a new system.

Resolves openbmc/docs#68

Change-Id: I16924b7991bc5e5a9111fdfd9a0bc18828ab4363
Signed-off-by: Andrew Geissler <geissonator@yahoo.com>