Update patch set 1
Patch Set 1:

(21 comments)

I made it halfway through this review, and I'm struggling a bit.  Most of this "design" comes down to

1. Entity-manager has some bugs.
2. Not all interfaces and use cases got documented upfront, and code is difficult to read.
3. Some daemons ignored the EM reactor design when they got built because of the technical barrier to entry for reading/writing config entries on dbus.
4. The reactor design needs better documented for things like the component add/remove event.
5. There exist reactor daemons that don't have accompanying EM config examples (ie peci-pcie, because most add-in-card type devices tend to be "secret").

Where I struggle is that there's efforts to fix some of these that are in review and clearly you have an interest.  If this design doc is asking for help in documenting and making things more consistent, I'm all for it.  If this design CCed the maintainers of the repos that it calls out and asked for design changes to be made, I'm all for it.  If this is asking the TOF to universally say "this is how it shall be" without discussion with the maintainers of those repos, I can't really get behind that.

Considering I'm the EM maintainer these days, and a lot of these complaints are things that I hadn't heard before, I worry that we haven't messaged this enough.  There will always be a class of things that are works in progress (ie peci-pcie).  If they do what you need, use them.  If they cause you problems, don't use them or push patches to fix them.  If you don't understand the design or think it's inconsistent, ask.

I'm really glad that you're using EM, and are working to improve it, I just struggle with how broad this document is, while still lacking critical understanding of a number of the use cases.  This is somewhat telling me that I need to write more docs for EM use cases, as well as start enforcing that new features get at least one example of use (ie peci-pcie) so that we don't have this misunderstanding.  FWIW, this has been on my list for a while now.  If you're interested in helping, always happy to have the help, let me know what you're interested in tackling and I can likely give you some pointers.

Will finish reviewing this doc tomorrow.

Patch-set: 1
CC: Gerrit User 1000153 <1000153@adbd0a64-1f21-4d83-b585-671fe73cb6e4>