enhance target-fail-monitoring for services
BMC systems are becoming more and more critical in todays data centers.
If a critical service within the BMC fails, it's paramount that
appropriate debug data be collected and the user of the system be
notified that the BMC is in an undefined state.

This new feature will allow a user to pass in a json file with systemd
service names that they wish this function to monitor. If a monitored
services goes into an error state (exhausted all retries and service has
been stopped) then the monitor service will create an error and collect
appropriate debug data.

Change-Id: I9ddc05387f23c80e490d72561397ab2850a05a3e
Signed-off-by: Andrew Geissler <geissonator@yahoo.com>