fail-boot: allow host to gracefully shutdown
Utilization of this feature in IBM manufacturing brought to light a
deficiency in this design. When the host firmware is booting and logs an
error that triggers this function, it would be ideal to allow that host
firmware the opportunity to gracefully shut themself down prior to
stopping them and moving the host state to quiesced.

This allows the host firmware to find and log any other relevant errors
(think something like verifying all dimms in the system). It's much
better to find all of these in one boot vs. needing to boot each time to
find each bad dimm.

This also allows host firmware a chance to properly write out any cached
data and handle any other relevant shutdown operations.

Change-Id: I99dd4a6afa2bf943eff87ef8f2fe670ebd264052
Signed-off-by: Andrew Geissler <geissonator@yahoo.com>