Add organizational section to the template
It was commented several times that we could significantly decrease our
time to repo creation if we called out in the design doc, whether or not
this feature requires a new repository, and who will maintain that
repository.  This commit adds those questions to the design template to
that effect.

Signed-off-by: Ed Tanous <edtanous@google.com>
Change-Id: I6972494d12d91b897780eb0469d7a7188472eca4