Make layer-guidelines a real markdown document
Add headers/anchors.  This facilitates easy copy & paste of links to
specific sections of the document.  Remove periods from new headers
because markdownlint doesn't allow that.  Reword the EXTRA_OEMAKE
guideline so it fits on a single line.

Add bolding and linebreaks around **why?** and **What should I do
instead?**.  The intent seems to have been to have these render on their
own line (bold and not H3 because markdownlint does not allow multiple
headings with the same content).

Add linebreaks in other places where the intent was for a new paragraph
to be rendered(I think?).

Change {autorev} to make it look more like it actually appears in the
code.

Signed-off-by: Brad Bishop <bradleyb@fuzziesquirrel.com>
Change-Id: Ib49d9e1079ff3a8b27c788ae34e1868a01cffc8f