Fix redfish cheat sheet firmware update
Hardcoding redfish paths is contrary to the Redfish standard.  While the
redfish cheat sheet hardcodes a number of paths, the firmware update
path is particularly hairy, as it's preventing us from effecting a
change that is correct per the standard, but because clients have copied
the incorrect cheat-sheet, requires openbmc to do code gymnastics to
keep some semblance of backward compatibility.

This commit fixes the documentation to call UpdateService, and pull out
the appropriate fields to then make a subsequent call to upload the
payload.

Signed-off-by: Ed Tanous <edtanous@google.com>
Change-Id: I5b822ec9f4495f1dcacd9973f31787ed23163870