RFC: clang-format: cpp: remove empty lines
This seems to be how most people code anyway (no metrics gathered just
my anecdotal observation), and with the setting in the other state,
whether or not extra lines are allowed appears ambiguous and can result
in comments and questions in review.  Take the hard-line to make it
explicit to avoid comments during review.

Change-Id: Icb75b3e11b5209a19c3ef6e92cf42d84a060af98
Signed-off-by: Brad Bishop <bradleyb@fuzziesquirrel.com>