Proposal for chassis blackout.target
Some OpenBMC systems have services they need run after a system blackout
occurs. This design change proposes a new target to coordinate these
services in order to alleviate discrepancies between system state or
services and their corresponding features.
Due to the unpredictability of a blackout, the system needs a mechanism
to acknowledge when one occurs, and also the means to assert proper
behavior within the system that reflects this changed state. With no
means to compensate for this issue, both physical and virtual features
on the OpenBMC server may exhibit improper functionality, resulting in
defects.

Signed-off-by: Corey Hardesty <corey.hardesty@icloud.com>
Change-Id: I7798b758a539bfae408ddf1d8422b144fcaf7a34