Update organizational section with maintainer cc
Lots of designs get submitted that seem to omit the maintainers from the
CC list, as well as designs that focus on only one part of the design
(sometimes dbus, sometimes redfish, sometimes IPMI) without considering
the whole system.

This commit adds a section to the design template for "repos that are
expected to be modified", so that we can make sure the required
reviewers are present on the design.  This also helps as a hint to
reviewers that they need to account for their whole design
(hw->dbus->user interface), not just a single piece, like dbus->user
interface or hw->dbus.

Signed-off-by: Ed Tanous <edtanous@google.com>
Change-Id: I74d34dafe7da0c83484df7afa96d3aef12efcc27