aggregation: phosphor-state-manager impacts and design
Aggregated systems contain multiple OpenBMC instances which must act
together as a single system.

The goal here is to capture the high level impacts of aggregation to
phosphor-state-manager and put some high level stakes in the ground on
how that will be handled. The expectation is that much will be found as
incremental patches start to find their way into upstream which may
results in some changes to this design.

Change-Id: I685b4c25655f39d1c6188a659d3f72cbc2b63422
Signed-off-by: Andrew Geissler <geissonator@yahoo.com>